// src/generated/models/Dga_module_typesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface Dga_module_typesBase {
  dga_module_typeid: string;
  dga_module_name: string;
}

export interface Dga_module_types extends Dga_module_typesBase {
  createdbyname?: string;
  createdon?: string;
  modifiedbyname?: string;
  modifiedon?: string;
  statecode?: number;
  statuscode?: number;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
}